Upgrade to Godot 4.0 and simplify
I simplified the class. Defining properties manually is no longer allowed,
now it needs to be done with `add_mandatory_property()` etc. Also all
properties are now internally arrays (previously it was mix of arrays
and strings), with type equal to TYPE_MAX for untyped variables. The
manual array definition was clunky anyway, so I doubt anyone used it.